Fix findings
