+++ Core/Core/Features/Inbox/ComposeMessage/Model/RecipientInteractor.swift
@@ -20,7 +20,7 @@ import Foundation
 import Combine
 
 protocol RecipientInteractor {
-    func getRecipients(by context: Context?, qualifier: GetSearchRecipientsRequest.ContextQualifier?) -> AnyPublisher<[Recipient], Never>
+    func getRecipients(by context: Context?, qualifier: ContextQualifier?) -> AnyPublisher<[Recipient], Never>
 }
 
 final class RecipientInteractorLive: RecipientInteractor {
@@ -29,7 +29,7 @@ final class RecipientInteractorLive: RecipientInteractor {
     private var subscriptions = Set<AnyCancellable>()
 
     // MARK: - Functions
-    func getRecipients(by context: Context?, qualifier: GetSearchRecipientsRequest.ContextQualifier? = nil) -> AnyPublisher<[Recipient], Never> {
+    func getRecipients(by context: Context?, qualifier: ContextQualifier? = nil) -> AnyPublisher<[Recipient], Never> {
 
         guard let context else {
             return Just([])
